# frontend/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Home Automation API",
    "description": "A FastAPI application for managing home automation routines and actions.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/routines/": {
      "get": {
        "tags": [
          "routines"
        ],
        "summary": "List Routines Handler",
        "description": "List all Routines ordered by creation date descending.\n\nArgs:\n    session: Injected database session.\n\nReturns:\n    List of Routines with their nested Actions.",
        "operationId": "list_routines_handler_routines__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoutineResponse"
                  },
                  "type": "array",
                  "title": "Response List Routines Handler Routines  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "routines"
        ],
        "summary": "Create Routine Handler",
        "description": "Create a new Routine.\n\nArgs:\n    payload: Validated creation payload.\n    session: Injected database session.\n    _user: Authenticated username (enforces JWT auth; value unused).\n\nReturns:\n    The created Routine with its nested Actions.",
        "operationId": "create_routine_handler_routines__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/routines/{routine_id}": {
      "get": {
        "tags": [
          "routines"
        ],
        "summary": "Get Routine Handler",
        "description": "Retrieve a single Routine by ID.\n\nArgs:\n    routine: Injected Routine instance (raises 404 if not found).\n\nReturns:\n    The requested Routine with its nested Actions.",
        "operationId": "get_routine_handler_routines__routine_id__get",
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "routines"
        ],
        "summary": "Update Routine Handler",
        "description": "Update an existing Routine.\n\nArgs:\n    routine: Injected Routine instance (raises 404 if not found).\n    payload: Fields to update (None values are ignored).\n    session: Injected database session.\n    _user: Authenticated username (enforces JWT auth; value unused).\n\nReturns:\n    The updated Routine with its nested Actions.",
        "operationId": "update_routine_handler_routines__routine_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "routines"
        ],
        "summary": "Delete Routine Handler",
        "description": "Delete a Routine and its associated Actions.\n\nArgs:\n    routine: Injected Routine instance (raises 404 if not found).\n    session: Injected database session.\n    _user: Authenticated username (enforces JWT auth; value unused).",
        "operationId": "delete_routine_handler_routines__routine_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/routines/{routine_id}/actions": {
      "get": {
        "tags": [
          "routines"
        ],
        "summary": "List Actions Handler",
        "description": "List all Actions for a Routine in position order.\n\nArgs:\n    routine: Injected Routine instance (raises 404 if not found).\n    session: Injected database session.\n\nReturns:\n    List of Actions ordered by position.",
        "operationId": "list_actions_handler_routines__routine_id__actions_get",
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActionResponse"
                  },
                  "title": "Response List Actions Handler Routines  Routine Id  Actions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "routines"
        ],
        "summary": "Create Action Handler",
        "description": "Create a new Action within a Routine.\n\nArgs:\n    routine: Injected Routine instance (raises 404 if not found).\n    payload: Validated creation payload.\n    session: Injected database session.\n    _user: Authenticated username (enforces JWT auth; value unused).\n\nReturns:\n    The created Action.",
        "operationId": "create_action_handler_routines__routine_id__actions_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/routines/{routine_id}/run": {
      "post": {
        "tags": [
          "routines"
        ],
        "summary": "Run Now Handler",
        "description": "Trigger an immediate execution of a Routine.\n\nInserts a RoutineExecution row and starts the engine in a daemon thread.\nReturns 409 if the routine is already running.\n\nArgs:\n    routine: Injected Routine instance (raises 404 if not found).\n    session: Injected database session.\n    _user: Authenticated username (enforces JWT auth; value unused).\n\nReturns:\n    RunResponse containing the new execution ID.\n\nRaises:\n    HTTPException: 409 if the Routine is already running.",
        "operationId": "run_now_handler_routines__routine_id__run_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/actions/{action_id}": {
      "put": {
        "tags": [
          "actions"
        ],
        "summary": "Update Action Handler",
        "description": "Update an existing Action.\n\nArgs:\n    action: Injected Action instance (raises 404 if not found).\n    payload: Fields to update (None values are ignored).\n    session: Injected database session.\n    _user: Authenticated username (enforces JWT auth; value unused).\n\nReturns:\n    The updated Action.",
        "operationId": "update_action_handler_actions__action_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Action Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "actions"
        ],
        "summary": "Delete Action Handler",
        "description": "Delete an Action and compact the position sequence.\n\nArgs:\n    action: Injected Action instance (raises 404 if not found).\n    session: Injected database session.\n    _user: Authenticated username (enforces JWT auth; value unused).",
        "operationId": "delete_action_handler_actions__action_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Action Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/executions/history": {
      "get": {
        "tags": [
          "executions"
        ],
        "summary": "History Handler",
        "description": "Return completed or failed Routine executions.\n\nArgs:\n    session: Injected database session.\n    limit: Maximum number of records to return (default 10).\n    routine_id: If provided, restrict results to this routine.\n\nReturns:\n    List of ExecutionResponse records, newest first.",
        "operationId": "history_handler_executions_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "routine_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Routine Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExecutionResponse"
                  },
                  "title": "Response History Handler Executions History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionCreate": {
        "properties": {
          "action_type": {
            "type": "string",
            "enum": [
              "sleep",
              "echo"
            ],
            "title": "Action Type"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          }
        },
        "type": "object",
        "required": [
          "action_type",
          "config"
        ],
        "title": "ActionCreate",
        "description": "Payload for creating a new Action within a Routine."
      },
      "ActionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "routine_id": {
            "type": "integer",
            "title": "Routine Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "action_type": {
            "type": "string",
            "title": "Action Type"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "id",
          "routine_id",
          "position",
          "action_type",
          "config"
        ],
        "title": "ActionResponse",
        "description": "Response DTO for an Action."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RoutineCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "schedule_type": {
            "type": "string",
            "enum": [
              "cron",
              "interval",
              "manual"
            ],
            "title": "Schedule Type"
          },
          "schedule_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Config"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "schedule_type"
        ],
        "title": "RoutineCreate",
        "description": "Payload for creating a new Routine."
      },
      "RoutineResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "schedule_type": {
            "type": "string",
            "title": "Schedule Type"
          },
          "schedule_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Config"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionResponse"
            },
            "type": "array",
            "title": "Actions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "schedule_type",
          "schedule_config",
          "is_active",
          "created_at"
        ],
        "title": "RoutineResponse",
        "description": "Response DTO for a Routine, including its nested Actions."
      },
      "RoutineUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "schedule_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "cron",
                  "interval",
                  "manual"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Type"
          },
          "schedule_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Config"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "RoutineUpdate",
        "description": "Payload for updating an existing Routine."
      },
      "RunResponse": {
        "properties": {
          "execution_id": {
            "type": "integer",
            "title": "Execution Id"
          }
        },
        "type": "object",
        "required": [
          "execution_id"
        ],
        "title": "RunResponse",
        "description": "Response returned immediately when a Routine execution is triggered."
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        }
      }
    }
  }
}